backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
huggingface-hub>=0.23.0
python-dotenv>=1.0.0

